Dagster_and_DBT_ELT/requirements.txt
dagit
dagster-postgres
dagster
dagster-dbt
sqlalchemy
psycopg2-binary
pandas
dbt-core
dbt-postgres
scipy